@@ -4,12 +4,12 @@ version: 0.4.4
 homepage: https://github.com/HomeXLabs/open-mail-app-flutter
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
-  flutter: ">=1.12.0"
+  sdk: ">=3.0.5 <4.0.0"
+  flutter: ">=3.3.0"
 
 dependencies:
-  platform: ^3.1.0
-  url_launcher: ^6.0.2
+  platform: ^3.1.5
+  url_launcher: ^6.3.0
   flutter:
     sdk: flutter
 
